[WIP] Initial commit of elementary Parts/Properties
This commit includes an initial implementation of custom Part
Properties, per the WyldCard documentation.

The current structure is a bit sloppy -- habit pushes me to just
implement this in term of simple properties on normal JS objects, or
even on the Part objects themselves directly as JS properties. Given
that we will have to add eventing to these specific types of Part
Properties when they change on their owning Part, this is a bad
idea. However, my current implementation here might not be the best
one.

Note that there are two "kinds" of Part Properties: Basic and
Dynamic. A Basic property is just one that stores a value and
retrieves it when asked / sets it when told. A Dynamic property
expects to be able to do some "computation" (look stuff up in its
owner, for example) before returning and/or setting a value. The
canonical example of the latter is getting a Part's "number". The
Number is an index of Parts of a certain type inside the owning part's
collection of child parts -- for example "button 3 of this
card". Ascertaining the correct index value is a function of the
PartsCollection and not the Property itself.

Therefore DynamicProperty instances expect to be initialized with
getter and setter functions whose first argument is always the Owner
Part instance, and who know how to look up whatever information they
need.

Note also that we have basic implementations of Stack, Card, and
Button in this commit, so that you can see where things might go.