Adding basic command handling, including delegation
Parts are now equipped to send, receive, and delegate
commands. Additionally, they each mainain a dictionary of handlers for
each named Command. These handlers, when called, will be bound to the
calling Part instance using `bind()`, to ensure that the `this`
keyword corresponds to the Part itself.

Any untrapped / passed command messages are "delegated". There is a
specified delegation/passing order in the system, which is this:
* Buttons and Fields delegate to their owner Card
* Cards delegate to their current Background
* Backgrounds delegate to their owner Stack
* Stacks delegate to "the system" (note: we don't have a concept of
this yet)

Delegation is implemented as the `delegateCmd` method. Each part is
responsible for overriding this method in its own specific
definition. For example, the current definition of Card overrides
`delegateCmd` to send the command to its current Background, etc.