Switching to more generalized message passing/delegation
The sending, receiving, and delegation of commands and functions is
now handled as part of a more generalized message passing design
between related Parts.

Messages are going to be basic JS objects, each of which has a 'type'
property specifying what kind it is. The only two the system expects
at this point are 'command' and 'function'.

For command and function type messages, we will probably expect
another property called arguments that contains an array of the
arguments to the command or function handler that will eventually be
called should the Part find one.