Adding property subscribers for Parts
We add the capability for any Part to keep track of objects that
'subscribe' to changes on any of the part's properties. This allows
other related JS objects -- for example, View objects on these Parts
where the Parts serve as models -- to update their visual
representations in the event that a property has changed on the model.

All Parts now maintain a private set of subscribers that can be
subscribed/unsubscribed to via the `addPropertySubscriber` and
`removePropertySubscriber` methods, respectively.

Now when a property on a Part is set, the Part's `propertyChanged`
method will be called with the name of the property that was changed
and the (new) value. The default implementation here is to use the
built-in message passing capability, creating a new message about the
property change and sending it to all subscribing objects. They can do
what they please with that information.