Adding first View example and Webpack config
This commit adds  Webpack config, which creates a small bundle in the
examples/dist folder that will then be used when importing the
wc1.html example.

The wc1 example is a proof-of-concept that a generic PartView can
load, then be given a generic Part instance as its model, and it will
correctly append its serialization state to the child script
element. Additionally, any subsequent calls to change properties on
the generic Part instance will end up triggering an update to that
script tag if a property has changed. This seems to be working for the moment.