[WIP] Working with new System object
The new System object is designed to be the 'broker' between the
Models (Parts) and Views (PartViews).

System is responsible for loading the page, deserializing any state
for models, attaching models to existing views, creating new Parts in
the sytem (new models), creating new Views in the system, and
serializing Part/Model state.

Note that this commit does not fully work yet. I need to rewrite a lot
of the View component logic and fix some cruft in the Part models.