Handler Additions (see comments below)
This commit adds the following:
* System handler for mouseUp, which does nothing. It swallows
the event, preventing a DNU error from being thrown;
* System handler for a command called saveHTML which prompts the user
with a save file dialog. The file being saved is the current HTML of
the page as has been modified by the running system. The idea is that
it can be reloaded and all of the saved state should deserialize and
continue, like a live snapshot;
* ButtonView now sends the mouseUp message to its Button model when
the DOM mouseup event is fired. You can test this by adding the
mouseUp handler for any particular button, or a card or stack in which
that button appears