[WIP] Adding ability to test custom elements and more
This commit adds the basic scaffolding for testing custom elements
using JSDOM (see below). The initial test file that implements this
capability as an example is an integration test of the binding between
DOM mouse events (specifically mouseenter) on a ButtonView and the
transformation of that event into a command message (mouseEnter) that
is properly handled by the system.

The mouseEnter test
-------------------
The new test file, aside from testing some initialization of the
system, will only test the mouseEnter message creation for the
moment. The behavior we expect is:
* If there is no object that 'traps' the mouseEnter message, the
System will receive it. However, unlike user-defined command messages,
we *do not* expect the System to throw a DNU -- mouse events should be
"swallowed" without disruption if no object traps them. For this we've
introduced the `shouldIgnore` property on command message
objects. When set to true, the System will not throw a DNU if it
receives this message as a last resort but also doesn't have a handler
for it.
* If the button itself has a mouseEnter handler, it should trap it.

The current test confirms this is working.

The Testing Setup
-----------------
There are several difficulties when it comes to testing web
components. First, we are accustomed to using JSDOM, which does not
have full capability for the moment (ie, we cannot test certain shadow
DOM features).

The bigger problem, however, is that our custom element definition
files expect a global object called HTMLElement to exist. Even test
files that load jsdom will run into an error when attempting to
import, either directly or indirectly, any of our custom element
modules. The solution is to force mocha to require a "preload" file
that creates a jsdom and sets mocha's global variables to what would
otherwise be global in the browser: window, document, and even
HTMLElement.

You can see this file in js/objects/tests/preload.js and you can also
see how it is required in the mocha command line by examining the
package.json file for the test command.

For the moment the basic test is working and it appears we can perform
non-shadow tests on our custom elements, including complex
integrations like eventing and message delegation.