Adding #goToNextCard and #goToPrevCard methods to StackView
This implements the first pass at basic card navigation in an active
stack. We implement the navigation on the StackView rather than the
model (a Stack instance), since the latter does not keep track of
which subpart Card model is the 'current' one -- this knowledge is
present in the DOM via the presence of the '.current-card' class on a
child st-card element.

We have also added (untested for the moment) a System method for
goToNextCard which performs the additional first lookup of the current
stack being viewed in the system, since the World/WorldStack has the
capability to hold several stacks at once (and displays one at a time,
like a Stack displays cards one at a time).

Included is a new test suite for StackView that tests nagivations
forwards and backwards in Stacks with three cards. All tests are passing.