Updating how tests work (with script)
The Issue
---------
For reasons explained elsewhere, we cannot reliably test webcomponents
when we are running Mocha over a directory and using the require on
the preload.js file to build a JSDOM (all tests end up sharing that
DOM, which is a disaster).

So instead, I'm adding a little bash script that finds any file
starting with test-* in the tests directory, and then runs that using
mocha and the appropriate require files (including preload). This ends
up being geologically slow, but all of the tests work as expected.

Perhaps there is a better way.