Adding webpack config for System.js and updating example
Because of how Ohm is bundled by its creators, we cannot fully use ES6
imports when adding our System.js to a web page. This means we need to
bundle that file.

This commit adds the webpack config and package command for creating
that bundle. Simply run `npm run build-dev` and the bundle will be
placed in `js/objects/build/devSystem.bundle.js`.

In addition, I have updated the system-example.html page to include an
initial WorldStack with all of its minimal part requirements. I did
this by "bootstrapping" the system, which entailed the following
steps:

1. Built the bundle
2. Opened the previous system-example.html file in a browser. This is
a "blank" file in that there are no parts or views yet, not even a
WorldStack
3. Ran this: `System.newModel('world', System)`, which created a
WorldStack, its view, an initial Stack, its view, and an initial Card
with its view.
4. I serialized the resulting page and saved it by sending this
message to the System:
        `{type: 'command', name: 'saveHTML', args: []}`
5. I then renamed the saved file to system-example.html, overwriting
the blank one.