[WIP] Adding Window Part and new System toolbox command
This commit adds the corresponding Window Part for the already-created
WindowView. The part only has a couple of properties for now, and the
view is only registering changes on the 'title' property. See the
class description for a bit more information.

We have also added a new System command called 'openToolbox' that is
kind of a visual test. For now, it configures a new Window with a
stack and opens it in the WorldStack. For now there are no buttons or
anything in the window but we can configure this later.

For now, sending the command manually as a message will open the
window in the WorldStack.