Refactoring preload.js so tests don't produce errors
Some tests are still failing, but there should not be any more
rudimentary errors. The key issue here was the loading of the JSDOM
and then loading the System file.

If you look at preload.js, you will see how this is working. System
has to be imported *after* the initial DOM document/window have been
created and attached to the global namespace. This is because System
adds a listener for the document's 'DOMContentLoaded' event and then
does its own initialization, like registering custom elements and
trying to create a WorldStack, if needed. These methods should not be
called twice as they will produce errors, so we do it all in the event
listener.

So all we need to have a working system with each test file is to call
preload before it.