Refactored flow of Model/View creation from System
This commit deals broadly with the problem of the System creating new
models and views. We have run into several issues with the current
setup that have made this process hard to track and difficult to
debug, and the design is not that great.

Here we implement a few big changes that hopefully simplify the matter
a bit. They also bring the Model => View flow more in line with what
one would expect.

One related immediate Issue was [Issue
10](https://github.com/UnitedLexCorp/SimpleTalk/issues/10), wherein
any model created by sending System the newModel message with an owner
attached wasn't actually putting the newly-created part model in the
parent/owner's subparts array (via calling `ownerPart.addPart()`).

I have now added this capability and it seems to work. The fix is in
the `newModel` method of System.
We might want to add further tests in the future. This gives is a
definitive tree of
models from which the System can produce and/or update View trees and
subtrees. It has immediate relevance to the changes below, but also to
a possible future refactor where we do more vdom-like stuff with
state, etc.

The `newModel` method on System was a mess. Now, instead of casing on
the type of model to create, we simply check the already-existing
registry of `availableParts` in the System. If there is a class there,
we use it to construct the model, otherwise we throw an Error.

This new version of `newView` looks up the model/part it will make a
view instance for. It checks to see if there is already a view for it
in the DOM and, if so, it simply returns the instance of that
element (**Note**: This means that, for the moment, System will not
handle multiple views for a single model, though we could easily
implement that in another way). If there is no view yet -- and this is
what has really changed -- System looks at the model itself, makes the
view, then calls `newView` recursively on any subparts of that model.

This ensures that System is keeping a 1-1 relationship between parts,
subparts, and the views for all of these things.

Previously, we were using the connectedCallback of StackView to see if
there was an existing Card or Background view and, in cases where
there was not, we sent the newModel message to System for those
parts. This was not the way to go for two reasons:
1. connectedCallback gets called at a different time than usually
expected;
2. We don't need to do this anymore -- System will create views
whenever subparts are available on a given model. Which brings us to
the related next change...

The Stack constructor now automatically creates its needed initial
Background and Card subparts. Now the System will know to create views
for those subparts automatically whenever creating a new
Stack/StackView combo, and the StackView doesn't need to know anything
about this.

Additionally, we were trying to set the current-card class inside of
the StackView's connectedCallback as well. This was again not the way
or place to do this, since we can connect a StackView to a parent
element without having any children attached to it yet. Instead, we
move that check to the CardView (see below)

We have added checking for current-card class status into the
connectedCallback of CardView. Here's why: this function will be
called whenever the CardView is connected to a parent element. The
parent element for a CardView will (for now) always be a StackView.

So we check the parent StackView to see if any elements have a class
.current-card. If there is one, we don't do anything. But if not, we
set the incoming CardView to have the .current-element class. This
ensures that the first card created on a StackView will automatically
be set to .current-card and no further hackiness is needed

I believe these changes will help us refactor some of the tests to get
them passing in a sane manner for the time being. They are also a
step towards a simpler system