Getting more tests to pass (and test refactoring)
== Basic Idea ==
We only need the preload.js file in cases where we are testing
something that requires the System.js interactivity. I have updated
the js-test.sh shells script to thus only require preload.js on test
files that end in *-preload.js. Right now there is only one, which is
the test-button-system.js file. I've renamed it to
test-button-system-with-preload.js and for now it is the sole test
file that will run with a System.js object.

This allows us to use a simpler basic JSDOM in other test files where
we only need the simplest functionality.

An example of this is the test-stack-navigation.js file, which really
doesn't need System.js at all (now that we've included the Model =>
View refactorings from this branch).

All of the tests that are not compiler related are now passing.