Update to serialization/deserialization and Toolbox example
== About ==
This commit is kind of big, in that it fixes a long-standing issue
whereby pages saved by sending the 'saveHTML' command to System could
not be deserialized into working Worlds. These should now work as
expected.

As a test case, and an example to show of what this branch is really
supposed to be about, I have generated an example Stack page and
serialized/saved it using the saveHTML command. This example is called
'toolbox-example.html' and is meant to show of the opening and use of
the toolbox.