Fixing deserialization process and updating example file
This commit seems to fix issue 15.

Now, during the deserialization process, if we encounter a serialized
model that has no view for itself represented in the saved HTML tree,
we DO NOT create a view for it and insert (this was producing errors).

Additionally, we have re-created the toolbox-example.html file, which
was itself somehow corrupted. It could not be used to save snapshots
that themselves would deserialize correctly when opened. This should
now be working.