Updating how PartView IDs are handled
= What =
Previously, PartView elements had a 1-1 mapping between their `id`
attribute and the id of the underlying Part model they are
showing. There were a couple of problems with this, most importantly
that such a design restricts the system from eventually having
multiple views for a single Part model.

= Solution =
This commit introduces the new `part-id` attribute on all PartView
elements. We use this instead of the normal `id` attribute. System has
been updated to look for these instead, as have a couple of tests.

All current tests are passing.

= WorldStack =
We have made an additional slightly related change, whereby WorldStack
will now always return 'world' as its id. We can change this behavior
in the future when we get around to implementing remote worlds or similar.