[WIP] Adding close button to Halos and updating Window Close
The close button in WindowViews will now simply call PartView's
default onHaloDelete() method. This method is also used by the Halo's
new white "close" button. The method will send the `removeModel`
message to the System -- that's it. The System handles removing the
model and views from the DOM.

Note that in the future, special PartViews can override this default
behavior if they want to.