[SEE-COMMENT] Fixing Layout problems in Toolbox Window
We were having an issue where the Toolbox window (or any window
really) was not adjusting its pane's dimensions to the underlying
Layout, even if this layout had contents like buttons.

The issue us that in a Stack/Card, Layout is a normal Part, and as
such will want to be `position:absolute`. With absolutely positioned
elements, however, the parent element will never "know" how to wrap it
-- the only options are how to deal with overflow.

This suggests that "layout" should be more fundamental to certain Part
kinds, rather than being a Part in itself. We need to think about how
to incorporate this.

In such a refactoring, instead of having a user-created Layout Part,
all Parts that display things can have internal layout properties, and
under the hood we have a special layout element in the shadow DOM that
is *not* a View (but rather a utility). This way Cards will have a
layout by default. Additionally, we can take the desired functionality
we want from the current Layout part and just make something called
Area or Group, which also has a layout under the hood but allows the
user to specify where it will be on the current card.