[WIP] Adding LayoutProperties mixin, with initial Card/CardView impl
We introduce a new mixin called LayoutProperties that, when called
inside of a given Part's constructor, will give the Part Model the
appropriate properties it needs to specify layouts.

We have implemented this in the constructor of Card, as well as added
appropriate change handling for those layout properties in
CardView. The Stylesheet has been updated to deal with the new layout
classes.

Note: we still have the Layout/LayoutView pair in this branch. The
next step will be the refactor those out into something like Area/Grouping/Container