Initial Experimental Drawing Webcomponent
This commit includes 3 experimental pure webcomponents that together
comprise a "drawing" system.

Note that none of this is connected to SimpleTalk yet.

The three components are:

* DrawingView (<st-drawing>) -- Contains a <canvas> in its shadow dom,
as well as an area to display tool buttons. Passes its drawing context
to any available tools and calls specific methods on those tools
during mouse events (if a tool is available and active).
* PencilTool (<pencil-tool>) -- When active, will draw rounded lines
of a specified width in its parent <st-drawing> element. You can
adjust the width by changing the 'width' html attribute.
* EraserTool (<eraser-tool>) -- When active, will erase any pixels in
its parent <st-drawing>'s canvas. You can set the size of the eraser
by changing the width property in the inspector.

This is an experiment in the play between real-dom nested elements and
how they can both have an effect on their parent (drawing or erasing
according to some specifications/attributes) and display themselves
within the shadow dom of the parent element (as toggle buttons that
activate or deactivate themselves). This represents a tightly-coupled
webcomponent system in which the elements understand how to interact
with each other.