Integrating experimental Drawing Part/View with image setting
We now have Drawing/DrawingView integrated into the System.

Note that each drawing action in the view, when complete, will try to
encode the contents of the canvas as a base64 png image. It will then
set the Drawing model's 'image' property to be that serialized image
data. The setting of this property is then seen by the view, which
clears the canvas and draws the image data to it.

This is how we can serialize the contents of the drawing easily. See
the serialization area for the drawing to see this in action.

Also note we have moved the files/directories around a bit, and added
a new button to the toolbox.