Adding 'mode' property to Drawing and adjusting halo resize
* Adding a 'mode' property to Drawing. When set to 'drawing', the
corresponding view(s) will show the drawing tools. It is set to
drawing by default. It maps to the view element's `mode=`
attribute. Any other value for that attribute will hide the drawing
tools.

* Adding an optional `onHaloResize` method call to views that have a
current Halo open on them. When the resizer is selected and is being
dragged, and the target element has this method available on it,
implementors can customize what resizing movement does. The method is
passed the event's movementX and movementY values.

In the case of the DrawingView, we use the information to update the
dimensions of the canvas and then restore the original contents to the
top left corner.