Adding template slots to Halo for custom buttons and Drawing button
This commit does the following:
* Adds named slots to the 4 button areas around the halo, wherein any
calling PartView can add its own custom buttons to the corresponding
area by referencing the slot name.
* Demonstrates this capability by having DrawingView add a new button
to the Halo's bottom row when the Halo is opened on it. This button
toggles the display of the drawing tools by toggling the model's
`mode` property. See DrawingView's `initCustomHaloButton()` to see how
it is constructed.

Also leaning up logs and eventing in various drawing tools