Fixing MouseLeave drawing issue
An occasional issue cropped up where when drawing with one tool,
keeping the mouse down, leaving the canvas, releasing the mouse
button, then re-entering the canvas would not reset the drawing
operations. This manifested itself in the appearance of not being able
to draw with one tool after selecting a first tool then performing the
above action.

We introduce a handler for mouseLeave which cleans up this problem.

It also changes the behavior slightly: if you are drawing and leave
the bounds of the drawing canvas, the drawing operation will cease
regardless of what you are doing with the mouse button.

Also removing unnecessary test file