Initial implementation of prop change handling in PartView
-- The Problem --
We need a universal and easy way for PartView and any of its
descendants to respond to changes on a given property of its
corresponding model.

As a recap, the normal flow is the following:
1. A View is created, and its model is set by setModel()
2. setModel() registers the view to be a "subscriber" to property
changes on the model. In other words, whenever a prop changes on the
model, the `propertyChanged` message will be sent to any objects that
are subscribed -- including the subscribed View.

This commit introduces changes in the base PartView that allow us to
handle step 2 above more fluidly in any subclasses. We do this by
implementing `receiveMessage` in this base class -- descendants should
not override it. For the moment, the only messages a view should be
receiving are precisely these property change types.

When a property change is received, PartView will see if there is a
handler (a function) registered for it in an internal
`propChangeHandlers` object that is created at construction. If there
is a handler, it is bound to the current this context and then called
with the new property value and the partId as the arguments.

We also supply descendants an easy method for registering their own
property change handlers, via the `onPropChange` method, which takes
the name of the prop as the first argument and then a handler function
as the second.

The idea is to have the base PartView implement property change
actions common to all PartViews. Subclasses / descendant views can
then decide how to react to specific property changes on their own
terms -- especially properties that are unique to that kind of view.

[BUGFIX] Fixing leftover method def that was bugging tests

[BUGFIX] Fixing error in EricFieldView that was bugging tests

[BUGFIX] Fixing message sending issue that was bugging tests