Adding tests for PartView property change handling
This commit includes tests for the pattern of attaching property
change handlers to PartView instances. In particular, it tests
onPropChange(), which is a convenience method for registering a change
handler for a given property name. We check to ensure three main
things for the moment:
1. The handler is bound to the correct property name and added to the
view's internal `propChangehandlers` dict;
2. The handler, when called, will have the correct 'this'
context (which should be the view itself);
3. The view reacts as the handler describes once the model updates the
given property