Adding basic variableName and associates rules to grammar
We are also adding rules for integerLiteral, floatLiteral, and
compound anyLiteral. These will be used for the most basic type of
variable assignment, which is to put a literal into a new variable.

In this commit we implement the most basic "put/into" command, which
takes an anyLiteral as the first argument and a variableName as the second