Adding compiler evaluate function and putInto implementation
-- Details
This commit actually does several things.

First, it implements a useful `evaluate()` function in the compiler
module. This function takes two arguments: an object that is the
result of a semantic parse() operation and a 'context' that we expect
to execute the evaluated part inside of. The context is going to end
up being a Part instance for the moment.

The idea is that, in most cases, we return whatever the semantics
already returns for the parse operation. However, in the case of
things like variables, we want to delay the lookup operation until the
message in which the variable is being set/read is sent. Therefore
`evaluate`, when called on a variableName part of the semantics, will
actually attempt to lookup the given variable's value and return that instead.

We also add to the base command compiler the setting of an empty
object called `_executionContext`. This is where we will put
handler-local variables and their values as they are executed within
such a handler. Note that the lifetime of such local variables is only
during the execution of the handler.

Now, when we call `resursivelySendMessages`, we first see if the
messages have an args array and, if so, we map each arg to the result
of `evaluate()` on the arg. As said, in most cases this will result in
the argument itself being returned unaltered.

This commit also implements, successfully, the `putInto` handler,
which is the result of the statement:
   `put <anyLiteral> into <variableName>`.

For the moment, this is the only way to set variables in the scripting
system.

Finally, I am implementing and adding tests for the variable,
integerLiteral, floatLiteral, and anyLiteral grammar/semantics.