Fixing Command Message arguments issue in tests
-- The Problem
The way a command message is translated into a bound command handler
function works as follows:
* We lookup a command handler on the part (a function) and grab it if
present;
* We run Compiler.evaluate() on all the args in the incoming Message,
saving them as evaluatedArgs
* We bind the `this` context to the found handler function
* We call the handler function by first spreading the Message args
list and then adding the Message senders array as the final argument.

The last step above means that we always need to be passing a strict
number of arguments in a given Command Message, even if some of them
ar undefined or otherwise empty values. Otherwise the bound handler
will not know the correct place in its own arguments list to read the
origin message's senders array, or will put it in the wrong place.

This is indeed what was occurring until this commit in the
test-system-add-model-with-preload tests. I have gone through and
updated the example Messages to include a full arguments array.

We might consider a different structure down the road. But for now, we
need to always pass the correct number of arguments with a given message.