Initial implementation of Interpreter & InterpreterNodes
-- About
This commit builds upon the work of adding "descriptors" to the
semantics. The idea behind descriptors was that they would be
lightweight objects describing some kind of delayed evaluation that
should take place whenever a scripted command handler's compiled
javascript function is applied with its arguments.

In instances like variable declaration and part querying, we want the
"lookups" to happen at runtime and not compile time. So we were
passing these descriptor objects as the arguments to compiled message
sends, and using a simple `evaluate` function inside of Compiler to
loop through argumenst before executing them (inside of System's
\#receiveCommand method).

It became clear that this is not really part of the compiler, but
rather an "interpreter". So we are transitioning "descriptors" to be
InterpreterNodes and adding a new Interpreter which knows how to
interact with the running system, and also how to evaluate any nodes
it is passed.

The key architecture decision at this point is to have the Interpreter
merely call `eval()`, which should be a method implemented on any
InterpreterNode descendant. `eval()` will be passed the Part instance
that serves as its execution context, so that the Interpreter can do
things like look up variable values, etc.