Adding hard-coded System "ask" command.
The ask command takes a string or other literal as an argument and
asks it as a question to the user. The user is presented with an input
field in which to supply their own answer.

The expectation is that the user's input will be implicitly inserted
into the "it" variable in any message handler, for later processing.

(Under the hood we use javascript's `prompt()` function)