Initial pass at implementing "it" implicit binding
We want the return value of any executed commands to be placed into
the "it" variable.

This commit attempts to do that, by updating two general things:
* Adding the return value of any command message send to the
_executionContext variable for "it" inside of Compiler's
`recursivelySendMessages()` call;
* Making sure that all message related methods (sendMessage,
receiveMessage, delegateMessage, etc) anywhere and everywhere in our
code actually return a value.

The latter point is the most sloppy of all of this. We will just need
to ensure we are always returning something, even in the case of many
forwarded message sends.