[BUGFIX] Fix for Issue #69
There was a bug by which the openToolbox and openWorldCatalog command
handlers were not receiving the correct arguments in the correct
order. The problem had to do with our use of the spread operator when
evaluating arguments in receiveCommand, followed by the mandatory
final argument of the message senders list.

We have now changed all command handler function signatures to expect
the senders list as the first argument, and to spread the interpreted
message arguments thereafter. This resolves the issue.

Note that we have updated the Compiler and other files to bind the
arguments as described above, as well as adjusting several internal
methods to System to be able to handle the new function signature
without breaking lots of code in places where these methods are called internally.