Initial pass at user-defined messages with arguments
-- Description
Until now, users could only define custom messages (and their
handlers) that didn't take any arguments. This commit is the first
phase in brinigng in this capability.

We have had to refactor some of the grammar / semantics in order to
make this work. Of chief importance is the introduction of initial
rules for Expression and Factor, per the HT manual.

I have also added a couple of test files that deal with these changes.