Implementation of Semantic-based Interpretation
-- About
This commit represents a wholesale replacement of the Compiler,
Interpreter, and InterpreterNodes setup.

Instead, we use a new operation on a configured version of an Ohm
semantics -- called `interpret()` -- to live-translate SimpleTalk code
into values, messages, and functions that the JS base of the
SimpleTalk system can understand.

-- Motivation
Putting all interpretation into the semantics this way has decreased
the amount of code and also the complexity in adding late-bound
capability to the scripting system. For example, it is now quite easy
to add variable lookups, part queries, etc., because this part of the
semantics itself has knowledge about the Part instance in which a
given handler is being invoked.

-- Implementation
See the `compile` message handler in System.js.
The function used to make the semantic actions for the `interpret()`
operation encloses the part instance and a reference to the System
object as variables. This is how the semantic actions are able to know
what Part they are operating on.

Aside from this, the basics work the same: the highest-level
`interpret()` operation -- which occurs on a MessageHandler rule --
creates a new JS function with the handler name and adds it to the
target Part's commandHandlers dictionary. The body of this function
has some of the semantics information in its scope, and it is here
where the Statements are interpreted and corresponding messages are
sent. This ensures that we are only interpreting when the command
itself is called for execution.

-- MessageHandler Arguments
This commit also implements the MessageHandler arguments, which
essentially map incoming JS function handler arguments to any named
parameters in the MessageHandler definition.