Fixing grammar definition of stringLiteral
-- Description
The previous definition of stringLiteral explicitly defined a set of
special characters that are allowed to be inside of the
quoteMarks. This is incorrect -- we want any character at all that is
not a lineTerminator.

-- Solution
The solution was to include a middle rule that describes 0 or more
instances of any character, so long as that character isn't a
lineTermination or another quoteMark.

We have added a couple of additional tests and they are passing