Updating compilation / semantic binding in System
Previously, we were adding a newly initialized semantics object to the
System itself whenever *any* part's script got compiled. This is
clearly not what we wanted to do. Instead, we now add this to the part
object.

Additionally, we are adding error handling for cases where a script
has incorrect grammar. The error is now sent as a (currently not
understood) error message to the System.