[BUGFIX] Fixing issue with messageName parsing
-- Description
We found an interesting issue whereby 'incorrect' groups of message
handler definitions were being parsed as if they were grammatically
correct.

For example:
```simpletalk
on doMyThing what
    answer what
on doMyThing

on click
    doMyThing "doing it"
end click
```
should not be correct, since there are two
MessageHandlerOpens (starting with "on") without the first one being
closed. It turns out ohm was seeing this whole script as one big
MessageHandler, starting with on doMyThing and ending with end click.

The culprit turned out to be the definition of messageName, which was
not correctly discounting keywords like "on". This means the inner
handler open lines were being interpreted as user-defined commands
rather than handler opens.

We have now fixes the definition. Additionally, we have fixed the
definition of objectId so that it no longer throws an error in the
newer version of ohm