Adding Conditional and control flow rules to grammar
-- About
This is an initial commit with grammar rules for various comparisons
and Conditional (expressions that evaluate to true or false) and
if-then-else control flow rules.

For the control flow, we currently have two general variants:
* IfThenInline -- things like `if <condition> then <command>` on a
single line. This option does not permit an else
* IfThenMultiline -- comes in two flavors:
    * `if <condition>
       then <command>`
    * `if <condition>
       then <command>
       else <command>`

-- Tests
This commit only includes tests for the basics of the conditional
grammar. More tests for the control flow grammar will appear in future
commits as I flesh it out