Adding Comparison/IfThen Interpreter Semantics plus tests
-- What
This commit finally adds functioning Ohm "interpret" semantics for
Conditionals and IfThen related rules. This allows us to finally
implement and test if then branching within SimpleTalk scripts.

We have also added an integration test file that simultaneously tests
some of the comparison semantics as well as the if then branching in
various forms (single line, multiline, etc).

Note that we made one minor adjustment to the main handler function
compilation portion of the interpreter semantics: it is now possible
for an interpreted Expression to return either a valid message for
sending (from a command) or null, the result of some false branch
being evaluated that does not result in a message to send. We now
ignore such intrepreted results, meaning we don't try to send 'null'
as a message.

All current and new tests are passing.