[WIP] Fixing nested 'exit repeat' in if-then statement. Tests broken
-- What
The following example was failing grammatically, and is still failing
in its implementation:
  ```
  on click
    put 0 into Counter
    repeat until Counter = 5
       answer Counter
       put (Counter + 1) into Counter
       if Counter >= 3 then exit repeat
    end repeat
end click
  ```
This has to do with how the if statement evaluates its 'then'
rule. Previously, we limited 'then' to only invoke actual
commands. Now any Statement can follow a 'then', and we have made the
'next repeat' and 'exit repeat' rules possible Statements.

The issue now is that in the interpreter we do not have a good way of
determining how some nested interpreted value -- say from an if-then
block -- returns something matching the 'exit repeat'/'next repeat'
rule. We therefore cannot reliably know when or how to break out of
the loop given the current implementation.

We now have both a grammatical test (passing) and an example
integration test. The latter is still not passing and we need to
further consider options for implementation. Might require rethinking
the whole looping construct for repeat structures.