Fixing Interpreter Repeat Implementation
-- What
By making rules like 'exit repeat' and 'next repeat' full
Statements (as they should be), we needed to re-design the flow of
various rule semantics inside of the interpreter.

This commit represents a working implementation, though we will add
enhancements in forthcoming commits -- in particular, domain specific
objects to distinguish between Statements that send messages and those
designed to control the flow of the program behind the scenes.

We have also added some integration tests for these to the test file
under the Misc category. They are all now passing.